Move json_history helper to chain/scripts/ folder
